coin and power as float

+++ src/main/java/uk/ac/ed/inf/powergrab/Drone.java
@@ -2,21 +2,21 @@ package uk.ac.ed.inf.powergrab;
 
 public class Drone extends Position {
 	
-	public double coin;
-	public double power;
+	public float coin;
+	public float power;
 	
-	public Drone(double coin, double power, double latitude, double longitude) {
+	public Drone(float coin, float power, double latitude, double longitude) {
 		super(latitude, longitude);
 		this.coin = coin;
 		this.power = power;
 	}
     
-	public void updateCoin (double coinIn) {
+	public void updateCoin (float coinIn) {
 		this.coin = this.coin + coinIn;
 		if (this.coin < 0) this.coin = 0; 
 	}
 	
-	public void updatePower (double powerIn) {
+	public void updatePower (float powerIn) {
 		this.power = this.power + powerIn;
 		if (this.power < 0) this.power = 0;
 	}
